backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy[asyncio]>=2.0.0
aiomysql>=0.2.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.25.0
pyyaml>=6.0
cryptography>=41.0.0
openai>=1.0.0
websockets>=12.0
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.0
python-jose[cryptography]>=3.3.0
